Preliminary: added node iterators
So far there is an unresolved dependency issue

Added an Iterator for each node to quickly process the stored
data. (only forward iteration relevant)
Added functionality for each node to adjust its size after an
insertion was performed.
Thereby, a LHC node can change the type to an AHC node if the
storage threshold is passed.

Elevated the node content struct into its own header file.

Added functionality in LHC to determin the longest prefix and
highest occupied addres. The former is used to determin the
threshold of changing to an AHC node and the latter is used
as end point for the related iterator.