Recursive deletion, fixed memory leaks
New:
-Differentiatied in the deletion behaviour. The delete AHC / LHC
will result in the removal of the information for this node but
not recurse to delete contained subnodes.
Therefore the method recursiveDelete was introduced.
If a node (LHC->AHC) changes its type it has to be removed without
deleting the referenced subnodes and suffixes.
The deletion of the entire tree will trigger a recursive deletion.
- Ensured that the NodeAddressContent can only have either a subnode
or a suffix reference at any time
- In LHC differentiate between a reference and a normal lookup. The
reference lookup can be used internally to directly manipulate the
values.

Memory (leaks):
- consequently does not use references to NodeAddressContent s any
more since LHC manages it differently than AHC and noone should be
responsible for differentiating when to delete. Only passing values
leads to copying (only pointers to subnodes and suffixes need to be
copied)
- Delete iterators (used as pointers) after usage
- Delete used entries in the plot utility

Misc:
- Renamed the property contained into exists (NodeAddressContent).
- Added some assertions.
- Further minor fixes.
- Ignore log-files like these from valgrind in git