Fixed memory leak for adjusted subnodes
When a node changed its format (LHC->AHC) it lead to a memory leak.
Now the parent node is responsible for removing outdated subnodes.

The LHC node must not delete an existing subnode if a new subnode
is inserted at the current location (valid scenario e.g. if the prefix
is split). Therefore it does not change any contained data except for
setting the suffix to NULL.

Corrected the recursive deletion in order to not delete memory twice.
Rewrote the assertions of the node class to get a stricter setting.