More compact and local AHC node data
Used a more cache efficient data layout for AHC nodes.
It does not store several vectors that need to be addressed
individually but instead used a struct to have all data in one
place.
(Exception: Does not apply for the suffixes which are still held
in a seperate vector to keep the data format tiny).

Also used a union to differentiate between a subnode and a suffix
entry.

Extended the SizeVisitor to take the data structures and their
formats into account.