Encapsulated bit handling, blockwise operations
Removed the static utility class to handle multi dimensional bitsets
and introduced the MultiDimBitset class.
Replaced all occurences with this class in order to have a central
point for changes. No implementation details need to be known from
outside.

Implemented the MultiDimBitset using blockwise operations on the
underlying boost::dynamic_bitset. I.e. load the stored unsigned long
and extract the relevant bits using bit operators. This is
supposed to be faster because bit iterators need to extract each bit
and convert it to the bool format.

Internally the MultiDimBitset stores bits in ascending order.
The PH-Tree addresses bits in descending order since the root node
contains the most significant bit. Therefore the addresses need to
be converted inside the bitset.
Instantly achieved a speed gain of 10% for the 3D benchmark.
For higher dimensions or longer values this should be more
segnificant.