Interleaving with block operators
Used the blockwise approach for handling the interleaving of bits
to produce an address.
Currently assumed that the dimension will never be bigger than 64.
Simply uses bit masks to extract the relevant bits from the block.
Since bits of the same depth (of one address) are stored consecutivly
the address can directly be used.
Attention: This changed the look of previous PH-Trees since the
addresses are now calculated in an ascending order but they are stored
in a descending order. Since the addresses have to be unique and
this property remains true the PH-Tree still works correctly.