Performance: minor tweaks + copy into &suffix
Used several minor performance optimizations:
- set several methods to const
- rearranged class members in order to get a smaller representation
(memo: C++ uses 64 bit blocks so int, int, long is smaller than
int, long, int in a struct)
(thereby reduced the PH-Tree size of the benchmark by 6%)
- used unions in the node contents to differentiate between a
subnode and a suffix
- used more initialization lists
- removed a constructor that required copying of the data structure
- made some members const
- removed one shift operation in the address extraction process
- added inline
- fixed minor memory leaks in the plot utility

If a suffix is inserted into a node the content struct is
created and an address to the suffix array is returned.
This can be used to populate the suffix.
Previously, unneccessary data structures were created and copied.

Overall speedup:
- lookup 4.4%
- insert 10%