In-memory size plotted per entry per dimension
Changed the plots for in-memory size to represent the number
of bits used per entry per dimension (divide by (n * d)).

Added output methods to all visitors in order to print their results
to the console.
Added a new visitor that checks how many bits are shared and how
big the saving is for these nodes. Therefore introduced a slow (!)
method that counts the number of existing contents (#suffixes + #subnodes)
of a node.

Added a minor performance change within the MultiDimBitset that uses
an inline method to calculate the bit length.