Used a template parameter for the dimension
Used the cardinality as a template parameter in all classes for
the following reasons:
- the AHC calculate the required array size at compile time
- the overall memory consumption is reduced since neither the nodes
nor the multi-dimensional bitvectors need to store the dim
- increased the performance as DIM is now a value known at compile-
time

However, the seperation between .h and .cpp files had to be
removed because class templates do not allow it.

Added minor performance improvements on the way such as using
initializer lists.