Minor performance tweaks
Used a single NodeAddressContent struct in the lookup and
insertion process in order to save costly recreation during
every iteration.
However, a helper method that has the same signature as the
previous lookup is available.

Removed the dim attribute from the entry and used the template
parameter instead.
Removed 'virtual' from several methods in AHC.
Added compiler ifndef to explicitly exclude the validation calls
after the insertion.

Resulted in a total speedup of 6% for insert and 8,6% for lookup.