Started to template the prefix blocks per node
In order to have a more local memory layout the prefix bit
blocks are now templated. Therefore each node has to store
an array of this exact size. The MultiDimBitset needs to be
rewritten in order to handle these arrays.
Suffixes should now be stored in blocks which are stored
in order in a global dynamic vector.

A new node was introduced that handles the prefix blocks
templating and hides this template parameter from external
access.

The insertion of a suffix should now be handled as follows:
1. calculate the remaining bits needed to store the suffix
(by tracking the index during the insertion)
2. allocate appropriate blocks in the global dynamic vector
3. extract the suffix from the entry into these blocks
using the MultiDimBitset
4. store a pointer to the start block of the suffix and the
id according to the node policy