Removed dynamic bitsets, templated prefix blocks
In order to have local data during the lookup all prefix bits
are now stored directly in the node. Therefore, a new node was
introduced which was templated with the number of 64-bit prefix
blocks. In these prefix blocks are all prefix bits stored.
The template parameted is determined during the creation process
of a new node.
Also the bitset of an entry was replaced with blocks.

Suffix blocks are not stored in a node any more but in a
superblock managed by the PH-Tree. Thereby, each node only needs
to store a pointer to the first suffix block. The number of actual
suffix blocks can be calculated during the lookup depending on the
number of processed bits until the suffix. Since superblocks store
all suffix blocks consecutively suffix blocks after the pointer
can directly be accessed.

The multi-dim bitset has been adapted to operate on these suffix,
entry and prefix blocks.

Notice that only for loading a subnode or for the final suffix start
block a cache fault should occure. Since blocks are stored consecutively
further suffix blocks should also be in the cache after the start block
was loaded.

Hence, when splitting a subnode prefix the subnode prefix blocks might
change and therefore a new TNode<DIM, PREF_BLOCKS> is created by
copying the contents.

The insert and lookup routine only operate on Node<DIM> and the actual
type TNode<DIM, PREF_BLOCKS> is dynamically infered. When a new node
is created a lookup for the correctly type has to be done.

Added the size of the superblocks to the plots.
Had to alter the visitor pattern because template methods cannot be
virtual.

Performance gains (in the benchmark):
- size: reduced by 42.3% because there is no unused space for suffixes
in the nodes
- insert: 33.5% faster because the lookups are faster and fewer objects
need to be allocated
- lookup: 22.5% faster because prefixes can directly be compared without
further memory accesses

Todo:
- currently realeased suffix blocks are not removed
- might not always need to copy when splitting a subnode prefix