Changed from RB-tree to own sorted array in LHC
Replaced the previous RB-Tree to an own implementation with a
sorted array. Insert might now take linear time as entries have to
be shifted. A lookup can be performed in log time with a binary
search.

Templated the LHC nodes with the number of maximum entries to
statically know the size of the required arrays. If the max size
is reached a new node will be allocated by copying all contents.

Benchmark:
As expected this slightly decreased the insert performance (-1%)
as contents have to be shifted and copied. The overall tree size
was slightly reduced by about 3%.