Integrated AHC flags into addresses
Integrated the two flags exists and hasSub into the lowest order
bits of the subnode references. This made the AHC node slightly
smaller and much faster because only one lookup is needed.
The lowest order bits are not used anyway for addressing.

Also added an LHC node of size 2 to the build utility because
every node starts with at most 2 entries.

Additionally, corrected the way the size is calculated for the
nodes. Previously the prefix arrays were calculated twice.

Benchmark:
- lookup: 8% improved
- insert: 4% improved
- size: 105MByte about 5% improved