Started to adapt range queries to the new layout
Started to adapt the previous rough implementation of range
queries to the new layout of prefixes/suffixes and nodes.

Encapsulated all node related contents such as the iterators
and the range masks in a struct. These node related contents are
managed on a stack.

The range queries not only need the visited nodes, but also the
addresses that lead to the visited nodes.

Added several new functions to the MultiDimBitset:
- remove highest bits within block:
This method should have the same effect as the existing
removeHighestBits with the difference that all changes effect
the given blocks directly.
- pushBackValue:
This method takes an interleaved address and inserts it at the
position specified by the parameters.
- pushBackBitset:
This method takes a multi dimensional bitset in block representation
and pushes the given number of bits to the other given bitset.
- compareSmallerEqual:
This method takes two multi dimensional bitsets in block representation
and compares them per dimension. That is, it extracts each dim-value
and compares it with the < and == operator to the extracted value of
the other bitset. In order to not really extract the interleaved
values a bitmask is applied that cancels out all bits that are not
part of the current dim-value and compares these results.

Adapted the debug example to contain more none 0 prefixes.