Started to put suffixes into references of nodes
If the suffix bits are smaller than the stored references there
is no need to allocate external memory for the suffix. Instead
it should be stored inside the references.

It might be problematic that the external suffix blocks are 8 byte
and the uintptr_t is 4 byte. Therefore, returning an address to
an internally stored suffix does not have the expected size.