Added a visitor for suffixes
Added a visitor that counts the suffix bits and suffix blocks.
Therefore, had to extend the accept method to also include the
nodes current index.

Realized that there are a lot of wasted bits because the blocks
are farely big. Also the benchmark produces long suffixes in
general but also holds a lot of nearly empty blocks that consume
more memory than the pointer needed to localize it externally.