More range querying
In case the last valid address of a node is reached a suffix
needs to be checked against the upper right range point.
(In the general node comparison resulted in no = comparison this
step can be skipped. I.e. all dimensions in which the = comparison
was true need to be checked.)

Corrected checking the current nodes range masks. Therefore
the lower left point needs to be compared with the current value
to which a 0 is added in all dimensions. The upper right point is
checked against a value with a 1 added to all dimensions.

Corrected the in place removal of bits.

Added a more efficient removal routine that only clear a hc address.
This touches at most 2 blocks and is the counterpart to the
setValue method.

Corrected the dimension wise comparison.