Continued range queries: plots, util, fixes
The range queries now work for the 3D runs but there are still
some errors for higher dimensions when multiple blocks are
accessed.

Rewrote the navigation routine within the PH-Tree for range
queries. It is now much more compact and can be used to
initially varify if the range is empty.

Added a method for multi dimensional bitsets that varifies that
a given bit range is unset (0). This is used in several
assertions in the range query.

Added new ranges in the simple example.

Added further comparison operators for the iterators in order
to varify the state.
Fixed the LHC iterator in case an address is set that is not
directly contained.

Extended the plot utility and the gnuplot files to include the
time for range queries (average millis per entry in the range).

Added a utility to run common functions on the ranges.