Suffixes up to 62 bits are now stored internally
All suffixes that hold up to 62 bits are now stored inside
the node as a reference. Previously they were stored externally
and the node would hold a pointer to this external storage.
Since 62 bits fit into 1 block, the reference to the external block
(64 bit) was bigger than the actually stored content.

The following layout was applied to the nodes:
- AHC: stores 62 bits of the suffix + 2 meta flags in the reference
- LHC: stores the 62 bits directly and holds the 2 meta flags in
the sorted lookup table for addresses

Adapted the lookup and the insertion to use internal suffixes.
Therefore, added a new flag to the container to determine if the
suffix is directly passed inside the container. If so, a reference
to the container can be returned instead of a reference to external
storage.

Added more assertions to the different insert methods.

Rewrote the lookup address method in the LHC in order to efficiently
handle the seperation of the relevant information in different blocks.