Added validation and fixes to range queries
Added a validation block to the next method of the range query.
1. lookup in the current node
2. lookup in the tree
3. range validation: lower left <= entry <= upper right

Rewrote the upper range check into a total range check. This
is neccessary but might be a point for optimization.

A shift by 64 bit results in a shift operation by 0. Therefore,
some multi-dimensional bitset methods had to be rewritten.