Corrected visitors, extended templates
The accept method previously did not correctly calculate the
index because the wrong prefix was used.

Added support for more prefix blocks. Also corrected the number
of required prefix blocks.

The random value generator can now handle bit widths for up to 64bit.

Improved the output of the suffix visitor