Interleaving in operations, spatial object support
Bounding volumes can now be inserted by using the lower left and
the upper right point of it. Lookups are also supported.
The values are simply reduced to a point. That is, if 3D space is
handled than the inserted spatial objects are 6D points. Thus, the
PH-Tree for 6D needs to be allocated.

Changed the plot utility to only use a vector of values. These
need to be interleaved in each operation. The old methods are
still available. This reduced the performance in the benchmark
but had to be done in order to compare to the Java solution.

Benchmark results:
- lookup: -46%
- insert: -34%