Faster range iteration, bitset fix
In case a node is fully contained an it was calculated that
all suffixes from this node are also within the range than all
subnodes can be iterated without calculating range masks or
validating suffixes.

Fixed the compareTo method for multi-dim bitsets that have an offset
in blocks. Previously, the dimension extraction mask was not
correctly calculated. Also the lowest block was off by one.