Faster (full) iteration over nodes
The range queries require a lot of accesses to the proper node
iterators. In case a node can be fully traversed some performance
optimizations can be applied:
- start iterator:
--> LHC: simply set to index 0 and load the address instead of
a full binary search
--> AHC: same as before
- end iterator:
--> both: can directly be set to the address 2^d as this is always
the end address
- access: since the access operator does not change the node
or the iterator it can be made constant
- removed the reachedEnd flag

The range query only needs one flag that determines whether the node
and all further contents are included in the range.