Added range query support for hyper rectangles
Range queries are now supported on hyper rectangles.
Added a simple example for normal rectangles in 2D and some
inserts, lookups and range queries.

The root node in the PH-Tree is now managed externally in the
insert routine in order to have a correct PH-Tree before the
end of the function (for validation).

Minor performance improvements.