Better rand generation, selectivity plot
Added a utility for generating random values using new C++ features
to ensure a correct distribution in a range.
Also ensured that 64 bit values are generated!

Added a plot that varies the selectivity of range queries and measures
the total time required for executing them. Therefore, a ratio
of selected space is given and placed at a random position in the
domain space.