logical/mem./perform./style fixes, inclusion query
Added inclusion queries (similar to intersection queries by require
spatial objects to be fully contained) and some samples in the
hypercube scenario.

Added the use case where dendrites (3D spatial objects) are indexed
using a PH-Tree and queried afterwards with axons to find
intersections.

New functionality for nodes:
- get the maximum prefix bits that the node can currently hold
- validate whether a given ID is in the current node by
linearly scanning its contents (slow but useful for validation)

Renamed specific range queries in the Ph-Tree to intersection and
inclusion queries.

The file utility can now read double values (without loosing (much)
precision). It currently relies on the number of decimals it
should use and a shift of the smallest (negative) value to the
positive value range.

Several suggested fixes from static code analysis:
- added a copy constructor for the PH-Tree
- all constructors with one parameter are now marked as 'explicit'
- removed unused parameters
- instead of pointer casting used reinterpret_cast

Added several further assertions.
Some fixes of memory leaks. (visited nodes, assertion visitor)
Range checks on multi-dim bitsets were adapted to only use
2 parameters that specify the range (the upper end can be
determined by the number of bits)
Restructured the compare method for bitsets and added a small
example.
Plotting text is flushed in order to follow the current work
in progress.

logical fixes:
- max. index in LHC iterators corrected
- the offset of hc address bits in block representation was not
calculated directly (should be the number of bits that cannot be
put into the current block)
- fixed removing the highest bits in case the parameter that
indicates the remaining bits is on the block boundary
- confused block and bit index in the removal of highest bits
- duplicate highest bits can now handle the case that the given
parameter is on the block boundary (multiple of 64 bit)
- clear value: off by one
- rewrote pushBackBitset to correctly work in all different
overlap scenarios and on block boundaries

performance:
- less calls to get the prefix length as it is now dragged along
in the insertion routine
- only duplicate remaining prefix if there is at least 1 bit left
to duplicate (same for the new prefix)
- only duplicate the longest common prefix if it is longer than
0 bits