Optimized comparing without LCP calculation
In the lookup routine the longest common prefix is not relevant
and therefore an optimized routine can be applied if two
interleaved bit sequences need to be compared. Also such sequence
blocks can be compared in linear - i.e. not in reversed order -
order from least to most significant bit which should be beneficial
for caching.

Unfortunately this did not really improve the lookup speed in the
benchmark.