idea: faster range queries
In order to speed up range queries the bits indicating whether
a value is equal in the current comparison in the dimension
corresponding to the bit inidex are stored on the stack.
When descending into lower levels of the tree there is no need
to compare values that are not equal in any dimension. Since
the bits are compared from most to least significant once a
value is smaller or greater than another it will not change
in deeper levels of the tree.

Storing all equality bits and providing them during the comparison
in order to skip certain dimensions or whole function calls.
Added a shortcut for suffix validation.

Currently still buggy.