Corrected local range checks in range queries
Previously only a single bit per dimension was used for local
comparisons. This does not work.
For local <= checks the current HC address bits and the
previous HC address bits are needed. Roughly implemented
the routine that calculates the mask ranges using previous
comparison results and local comparison results. If there
was no prefix only the address can be extracted. If several
bits need to be compared the previous function is used but
not from the MSB bit down but from the previously compared
index.

The code was not yet tested!