Minor fixes and more assertions for range queries
Fixed two issues for local comparisons in range queries:
- if a prefix is included the previous DIM bits, all prefix bits
and the current DIM bits need to be compared
- added an upper mask for the <= comparison

Added a new block which verifies that the locally combined
comparison is equal to the global comparison.

Added a new example which is a fully filled binary tree with
no prefix in the first level, a 0 prefix in the second level and
a 1 prefix in the last level.