Added local operators for range queries
Encapsulated the logic for local checks in the range queries.
Basically there are 3 different operations (all are
individually implemented for upper and lower checks):
a) compare a single address locally with the current and previous
boundary value
b) compare a range (n addresses) locally by overlapping with
the prevous value (1 address) -> compare n+1 addresses
c) connect a local comparison (single or range) with the previous
comparison to make a valid for all values compared so far.

Used these operators when validating the prefix and when
calculating the range masks.

TODO:
- use for suffix check
- currently double work when a prefix is valid: instead validate
prefix and pass comparison to range mask calculation