All local comparisons in range queries
Used the local comparison modules previously introduced to
do local checks for suffixes and prefixes.
Restructured the routine that descends into a new node:
- first :  the prefix is locally compared
- second:  depending on the result
-- a) valid prefix   | connect the local comparison to the previous
comparison results in order to pass it on. Extract the last compared
HC address and pass it on too. Thereafter, only need to compare
the range for the current node without extracting any values.
-- b) invalid prefix | abort the procedure of stepping into this
node