Only compare dimensions neccessary in range query
Since each dimension is compared individually many of these
comparisons can be skipped in the <= comparisons.
For the lower comparison no bits need to be extracted if
lower range < current is already true for that dimension.
For the upper comparison current < upper range has to hold.

This slightly improved the performance.