Storing suffixes in dynamic leaf storage per node
Instead of placing all suffixes in global super blocks that
are difficult to manage and that might degrade over time
all suffixes of one node are now stored locally in a memory chunk.

These memory chunks are templated and buffered similar to nodes.
The compilation and storage overhead of templating a node together
with the stored suffixes it to hight. Therefore, the suffix storage
was decoupled from the node.

Nodes do not store references to suffixes any more. Instead, they
store the index of the start block in the corresponding suffix
chunk. The reference can be resolved from the array reference
and the offset/index. If the suffix chunk is enlarged or shrinked
there is no need to update the indices whereas pointers would
need to be updated. To enable different behaviour in the
iterators resolving the indices to references can be enabled
(default) or disabled dynamically.

Benefits of the new layout:
- better memory management because there are no more unreachable
suffixes. In case a suffix needs to be removed it is swapped with
the last suffix and deleted afterwards. The reference to the
swapped suffix is updated afterwards.
- better caching: If a node needs to be traversed the suffix
chunk is loaded into the cache for the first access. Since all
of the suffixes of a node are stored in the same chunk all further
accesses can be served from the cache

Todo:
- store IDs/keys in these chunks too
- find better thresholds for the suffix chunks to reduce unused
space
- measure performance / memory gain