Store IDs/keys and suffix indices together
changed the LHC layout by removing the ID array. It wasted memory
for all subnode references as they do not have any associated ID.
Now used the following format to store an ID together with the
suffix inside the references array:
---------------------------------------------------
|   ID (32)     | suffix index / suffix bits (32) |
---------------------------------------------------

This limits the number of suffix bits that can be stored internally
to 32 but because this is a rare case for high dimensions it is a
valid change. This reduced the (rand) benchmark memory size by
several MB (about 10% smaller now).