Special pointers, entry buffer, bulk insertion
Additional new way of storing information in LHC/AHC:
- Special pointers (uintptr) can be inserted into nodes and
retrieved using the lookup. The size of the information wrapper
did not grow as the special pointer is stored in the union and
the flag that indicates the special pointer does not exceed the
64 bit alignment.
- Flagging special pointers:
-- LHC: 00 indicates a special pointer. Had to change the way
the pointers are interpreted to a format similiar to AHC.
Previously each bit indicated a flag whereas now the 4 combinations
of the 2 bits represent the 4 cases (subnode, suffix, internal
suffix, special pointer)
-- AHC: 00 is already used for representing empty references.
However, if the reference != 0 a special pointer can be stored and
clearly be differentiated from reference == 0.

Added a bulk insertion method that works similar to the normal
insertion with the following changes:
1) A suffix is not split. Instead a buffer is created and inserted
as a special pointer.
2) When a special pointer is found the buffer is resolved and the
new entry is copied into the buffer. If the buffer overflows it
is flushed by spilling its contents into a subtree.
3) The buffers always need to now the current reference to the
node they are attached to because:
4) In the end all buffers are emptied.

The EntryBuffer works as follows:
1) Initiate the buffer by indicating the node, the HC address and
the length of the suffix.
2) Copy the suffix into the buffer.
3) Each newly inserted entry calculates its longest common prefix
with the suffix and updates the overall longest common prefix.
4) If the buffer is full or the LCP is 0 the buffer should be
flushed.
5) Flushing the buffer
5.1) Sort the entries (insertion sort because small set) by their
LCP
5.2) Iteratively, create nodes of the correct size and insert
suffixes, prefixes and subnodes.

Current properties:
a) All entries with a unique LCP are going to be a suffix
b) The maintained suffix of the buffer is also going to end up
as a suffix
c) All entries with the same LCP might conflict (if they have
the same HC address) and end up in a subnode