Corrected top-down entry buffer flushing
Moved to lookup method that creates a new NodeAddressContent object
on the stack up in the hierarchy for easier and more general access.

Enabled collecting and updating the entry buffers by introducing
a general TEntryBuffer that only maintains a reference to the
original node. Thereby, it can easily be updated without the proper
template parameters (cf. NodeUtil).

The bulk loading routine finalizes the tree by alternating the
collections of filled entry buffers to flush. The flush function
may create new buffers to be flushed in the currently inactive
collection.

Extended the flush routine of the buffer to:
1. correctly calculate the number of suffixes and subnodes
required in the node
2. creating new buffers for overlapping entries
3. insert the buffers / suffixes into the correct node