Entry buffers with LCP matrix
Reimplemented the insertion and flushing in/of buffers. Instead
of comparing to the first entry only a matrix of longest
common prefix lengths is stored. This matrix is used to calculate
the subtree. With this new approach flushing does not introduce
new buffers and all inserted entries end up as a suffix.
The algorithm works roughly as follows:

1. per row find the current highest LCP and calculate the
number of suffixes and subnodes O(N*N)
2. per row that has the global max LCP create a new node with
a prefix depending of length max[row] - (1 + nextMax[row]),
insert each column with max[col] == max LCP and store the
new node for each inserted column. Cancel all but the
first matching column. O(N * N)
3. repeat until there are now columns / rows left O(log N).

Added a printable version of the LCP table during the algorithm
for debugging and visualization purposes.

When swapping in a buffer for a suffix the suffix storage can be
shrinked.
Only requires a single set of buffers since a flush does not
introduce new ones.

Created new facility methods for multi-dimensional bitsets:
- duplicate bits: works as duplicateHighestBits but ensures that
the result is correct even if there are bits set higher than the
given number of bits
- duplicate lowest bits (block aligned): copies all bits from one
reference to another from LSB up. It assumes that both references
are aligned and therefore works much faster than the other
duplication methods.

Extended the plot utility to perform bulk loading measurements.
Also the gnuplot files are now sorted by an index provided in the
first row using the external program 'sort'.
Added default constructor to Entry to create an array in the buffers.
Added a bulk loading example in the main method.
Corrected several accurances of checks with special pointers.
Added counters for new bulk loading function calls.
Refactored -DPRINT output.