Skipping empty columns during flush
Instead of checking whether each row is empty and than continuing
to the next several columns can be skipped.

Added an additional array that stores the offset to the next
valid column. Since each row is iterated from lowest to highest
column only the first invalid column in a streak needs to store the
correct offset. This trick is applied during the insertion and during
the calculation of the max LCP values. It reduces the required loop
iterations it both the for loop over the rows and the for loop over
the columns.

example (x is an empty column, X is the newly emptied one)
_ _ _ x x x _ _ x x _
      ----->    --->

_ _ _ x x x X _ x x _
      ------->  --->

_ _ _ x x x x X x x _
      ------------->