Storing symmetric LCP matrix
Instead of storing the hole LCP matrix only the lower half
(every value on the diagonal and under it) needs to be kept.
Adapted the storage and the access.
Storing the lower and not the upper half because the rows are
iterated in a linear fashion so all LCP values of one row should
be stored consecutively.

Also added a counter for buffer insertions.