Pooling for entry buffers
Since the buffer approach introduces many buffers in the tree
of which many might never be flushed until the end a pooling
system was introduced that is also flushed if it overflows.

EntryBuffers are now stored in a loop free and fragmentation
free memory pool with two basic operations:
- allocate: returns a pointer to an unused buffer and refreshes
the list of empty fields (that is stored within the buffers)
- deallocate: frees the given pointer.

The memory pool has a fixed size and can hence overflow. Instead
of creating a new one and introducing memory and computational
overhead the entire pool is flushed by invoking the flush operation
on each valid internal buffer. Thereby, outdated buffers are freed
again.
The concept of the memory pool was inspired by the paper
"Fast Efficient Fixed-Size Memory Pool" and extended afterwards.

Had to extend the clear method of the entry buffers to be able to
reuse objects. The bit fields were not previously cleared.

The entry buffers do not require an external storage of used
buffer pointers. Instead the used pointers are implicitly stored
within the pool as the set of entries that is not in the free
list. These entries can be retrieved doing the following:
1. Iterate over the free list and set all nextIndex values to -1
(Thereby, removing the free list) --> O(n)
2. Iterate over the remaining entries by flushing only those buffers
that have entries stored (nextIndex > 1) --> O(n)

As a third step the pool can be reset and the insertions start
again.