Implemented optimistic locking for bulk insertion
Implemented the optimistic locking strategy for bulk insertions
which works as follows:
1. when a node is read a read lock needs to be acquired (only up
to two read locks are held at any time)
2. when the node needs to be modified the read lock is upgraded
and than the thread waits until all present readers are done
3. once all readers are done the thread exclusively modifies the
node
x) if any lock acquisition fails the thread restarts.

Added a thread pool that sequentially processes all the data to
be inserted. During creation it builds the biggest possible
root node and during deletion it shrinks it to what is actually
needed. This is done in order to prevent synchronization on the
root node of the PH-Tree.

There are better processing schedules than the sequential one
as it requires a synchonized index variable.

In order to prevent memory reclamation issues no nodes of the
tree are deleted! This needs to be fixed.