Thread pool processing in predefined chunks
Instead of processing the entries in the thread pool from lowest
index to highest index in a sequential manner via an atomic shared
index each thread is assigned to a unique range of the data.
Since HBP data is structured this tries to avoid changes in the
same area of the tree and hence less congestion / restarts.

Also corrected the calculation of the parallel efficiency and added
the throughput.

Created a plot.