Insertion schemes, fewer locks, boost fix
The previous locking error was in the implementation of the
shared_mutex in boost. By uncommenting the two manual lk.unlock()
calls in unlock_shared() the optimistic locking works. A ticked
was filed for the boost library to fix this.

Added three different insert schemes for the insertion thread pool
and adapted the gnuplot file to print them.

Added a method to first check whether inserting a new value in a
node would require a copy. After that check only those nodes can
be locked for writing that are actually needed.