Parallel bulk load with buffers
Extended the previous implementation of parallel bulk insertion
with optimistic locking to incorporate the entry buffers.
Therefore had to synchronize the buffers and the pools.

1) Buffers:
- to access and insert into a buffer only read permission on the
parent node is required
- the insertion is synchronized via an atomic index counter
- busy waiting to compare to a previous value that is not finished
- once the counter exceeds the capacity of the buffer it is an
indicator for the congestion
- before inserting into a buffer it is emptied if it is full
and the recurive search continues from the current node (emptying
the buffer requires write permission on the parent node so there
is no need to synchronize this operation)

2) Pools:
- currently each thread has its own pool and once any pool is filled
a flushing phase is introduced through barriers
- Each thread then flushes its own pool without acquiring any
locks
- allocation and deallocation are synchronized through a unique lock

Slighlty changed the way pools are flushed. The pool now has the
responsibility to manage the valid buffers and to initiate their
flushing.