One barrier for flushing, compare parallel algos
Only one barrier is needed to synchronize the flushing phase.

Better output for parallel insertion performance. It now
compares the optimistic locking with the version that uses
buffers. Also two different orders are compared. Text formatting.

Removed time-consuming assertions from the buffer pool that
are also catched in the full deallocation.

Disabled the restart counters as they require synchronization
and are therefore unsuitable for comparing to the optimistic
locking approach.