Thread-safe edge updates in LHC
Previously the update of an edge (e.g. buffer -> subnode)
required to modify the map that sequentially stored the contained
HC addresses and 2 bit flags that indicate the type. Since
several threads might update bit flags that are stored in the same
block (64 bit) there might be race conditions.

Changed to the format that is already present in AHC nodes:
storing the 2 bit flags together with the reference. This does
not limit the current behaviour of the trees since the 30 bit
constrain for directly stored suffix bits and the suffix index
was already present due to the AHC node (64 bits = 32 bit ID +
30 bit suffix / suffix start block index + 2 bit flags).

Adapted the iterator.